Updated the LockService to allow for acquiring locks on resources.
opensearch_job_scheduler_lock.json - changed to include fields for the resource lock
LockModel - changed to include constructors for a resource lock
SchemaUtils - created to have functions that allow for LockService to check whether the cluster schema for the lock index is up to date.
LockService - allows for the acquisition of resource locks and also makes calls to use SchemaUtils before writing to the lock index.
LockServiceIT - extends tests to make sure that resource locks are tested as rigorously as job locks.